pyproject.toml
[project]
name = "greptile-mcp"
version = "0.1.0"
description = "MCP server for code search and querying with Greptile API"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Greptile Team" },
]
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
select = ["E", "W", "F", "I", "UP", "ASYNC"]
target-version = "py312" 